Add preliminary support for Windows
