+++ x/ref/examples/tunnel/vsh/main_posix.go
@@ -1,0 +1,13 @@
+// +build !windows
+
+package main
+
+import (
+	"os"
+	"os/signal"
+	"syscall"
+)
+
+func signalWindowSizeChange(winch chan os.Signal) {
+	signal.Notify(winch, syscall.SIGWINCH)
+}
