+++ x/ref/examples/tunnel/vsh/main_windows.go
@@ -1,0 +1,8 @@
+// +build windows
+
+package main
+
+import "os"
+
+func signalWindowSizeChange(winch chan os.Signal) {
+}
