+++ x/ref/services/agent/internal/launcher/launcher.go
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// +build !windows
+
 // Package launcher contains utilities to launch v23agentd.
 package launcher
