+++ x/ref/services/agent/internal/lockutil/version1_windows.go
@@ -2,10 +2,5 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// +build windows
 package lockutil
-
-const (
-	systemIDLabel = "SYSTEM ID"
-	pidLabel      = "PID"
-	unknownID     = "UNKNOWN"
-)
